scripts/requirements.txt
kafka-python-ng==2.2.3
redis==5.0.8
faker==25.0.0
